Updated Kilo
Added a texture for weightless and changed it so it uses an umbrella. I also added a new move called Kick Off Jump which launches you into the air. Finally I removed a bit of code which makes all umbrellas slow your descent as I think it makes no sense without the kilo